+++ projects/egeo/src/lib/st-dynamic-table/spec/st-dynamic-table.component.spec.ts
@@ -11,47 +11,90 @@
 import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
+import { cloneDeep as _cloneDeep } from 'lodash';
+import { JSONSchema4 } from 'json-schema';
 
-import { Order, ORDER_TYPE } from './shared/order';
-import { StPopoverFilterComponent } from './shared/st-popover-filter/st-popover-filter.component';
-import { StTableComponent } from './st-table.component';
-import { StTableHeader } from './shared/table-header.interface';
-import { StCheckboxModule } from '../st-checkbox/st-checkbox.module';
-import { StPopOverModule } from '../st-pop-over/st-pop-over.module';
+import { StDynamicTableComponent } from '../st-dynamic-table.component';
+import { Order, ORDER_TYPE, StCheckboxComponent, StCheckboxModule, StTableComponent, StTableHeader, StTableModule } from '../..';
 
-let fixture: ComponentFixture<StTableComponent>;
-let component: StTableComponent;
-let fakeFields: StTableHeader[] = [
-   { id: 'id', label: 'ID', sortable: true },
-   {
-      id: 'name',
-      label: 'Name',
-      sortable: false
-   },
-   { id: 'lastName', label: 'Last name' },
-   { id: 'phone', label: 'Phone', sortable: true },
-   { id: 'group', label: 'Group', sortable: true, filterable: true, filters: { title: 'test', filterConfig: [{ id: 0, name: '1111' }] } }
-];
+let fixture: ComponentFixture<StDynamicTableComponent>;
+let component: StDynamicTableComponent;
 
-describe('StTableComponent', () => {
+
+describe('StDynamicTableComponent', () => {
+   const jsonSchema: JSONSchema4 = {
+      '$schema': 'http://json-schema.org/schema#',
+      'title': 'Executor Table',
+      'type': 'object',
+      'description': 'Spark executor properties',
+      'optional': true,
+      'properties': {
+         'cores': {
+            'title': 'Cores',
+            'description': 'Number of per spark`s executor',
+            'type': 'integer',
+            'default': 4,
+            'minimumExclusive': 0,
+            'maximum': 8
+         },
+         'memory': {
+            'title': 'Memory',
+            'description': 'Spark executor memory',
+            'type': 'integer',
+            'default': 512,
+            'minimum': 512,
+            'maximum': 4096
+         },
+         'home': {
+            'title': 'Executor home',
+            'description': 'Set the directory in which Spark is installed on the executors in Mesos.',
+            'type': 'string',
+            'default': '/opt/spark/dist',
+            '$ref': '/root/homes/executor',
+            'enum': ['/root', '/desktop', '/document']
+         },
+         'subexecutor': {
+            'title': 'Sub executor',
+            'type': 'object',
+            'description': 'Spark executor properties',
+            'properties': {
+               'subcores': {
+                  'title': 'Sub cores',
+                  'type': 'integer'
+               },
+               'submemory': {
+                  'title': 'Sub memory',
+                  'type': 'integer'
+               }
+            }
+         }
+      }
+   };
+   const items: { key: string, value: any }[] = [];
 
    beforeEach(async(() => {
       TestBed.configureTestingModule({
-         imports: [CommonModule, RouterTestingModule, StCheckboxModule, StPopOverModule],
-         declarations: [StPopoverFilterComponent, StTableComponent]
+         imports: [CommonModule, StTableModule, StCheckboxModule],
+         declarations: [StDynamicTableComponent]
       })
       // remove this block when the issue #12313 of Angular is fixed
+         .overrideComponent(StDynamicTableComponent, {
+            set: { changeDetection: ChangeDetectionStrategy.Default }
+         })
          .overrideComponent(StTableComponent, {
+            set: { changeDetection: ChangeDetectionStrategy.Default }
+         })
+         .overrideComponent(StCheckboxComponent, {
             set: { changeDetection: ChangeDetectionStrategy.Default }
          })
          .compileComponents();  // compile template and css
    }));
 
    beforeEach(() => {
-      fixture = TestBed.createComponent(StTableComponent);
+      fixture = TestBed.createComponent(StDynamicTableComponent);
       component = fixture.componentInstance;
-      component.fields = fakeFields;
+      component.jsonSchema = jsonSchema;
+      component.items = items;
       component.qaTag = 'fake qa tag';
    });
 
@@ -71,30 +114,65 @@ describe('StTableComponent', () => {
          expect(component.hasHoverMenu).toBeFalsy();
       });
    });
+   describe('It should be able to apply some styles from outside', () => {
+      it('If uiDefinitions input is introduced and there is a field with styles properties, there are added to the right field', () => {
+         component.uiDefinitions = {
+            cores: {
+               styles: {
+                  width: '25%'
+               }
+            },
+            home: {
+               styles: {
+                  color: '#DF2935'
+               }
+            }
+         };
 
-   it('if table is sortable but there are fields not sortable, sortable class is only added to sortable fields', () => {
-      component.sortable = true;
-      fixture.detectChanges();
-      let headerItems: HTMLTableHeaderCellElement[] = fixture.nativeElement.querySelectorAll('.st-table__header-item');
+         fixture.detectChanges();
 
+         const rows = fixture.nativeElement.querySelectorAll('tbody tr');
+         for (let i = 0; rows.length; ++i) {
+            const cells = rows[i].nativeElement.querySelectorAll('td');
 
-      expect(headerItems[0].classList).toContain('st-table__header-item--sortable');
-      expect(headerItems[1].classList).not.toContain('st-table__header-item--sortable');
-      expect(headerItems[2].classList).toContain('st-table__header-item--sortable');
-      expect(headerItems[3].classList).toContain('st-table__header-item--sortable');
-      expect(headerItems[4].classList).toContain('st-table__header-item--sortable');
-   });
+            expect(cells[0].styles.width).toEqual('25%');
+            expect(cells[2].styles.color).toEqual('#DF2935');
+         }
+      });
 
-   it('if table is not sortable but there are fields sortable, arrow is displayed for sortable fields', () => {
-      component.sortable = false;
-      fixture.detectChanges();
-      let headerItems: HTMLTableHeaderCellElement[] = fixture.nativeElement.querySelectorAll('.st-table__header-item');
+      it('if table is sortable but there are fields not sortable, sortable class is only added to sortable fields', () => {
+         component.sortable = true;
+         component.uiDefinitions = {
+            memory: {
+               sortable: false
+            }
+         };
+         component.jsonSchema = jsonSchema;
+         fixture.detectChanges();
 
-      expect(headerItems[0].classList).toContain('st-table__header-item--sortable');
-      expect(headerItems[1].classList).not.toContain('st-table__header-item--sortable');
-      expect(headerItems[2].classList).not.toContain('st-table__header-item--sortable');
-      expect(headerItems[3].classList).toContain('st-table__header-item--sortable');
-      expect(headerItems[4].classList).toContain('st-table__header-item--sortable');
+         let headerItems: HTMLTableHeaderCellElement[] = fixture.nativeElement.querySelectorAll('.st-table__header-item');
+
+         expect(headerItems[0].classList).toContain('st-table__header-item--sortable');
+         expect(headerItems[1].classList).not.toContain('st-table__header-item--sortable');
+         expect(headerItems[2].classList).toContain('st-table__header-item--sortable');
+
+      });
+
+      it('if table is not sortable but there are fields sortable, arrow is displayed for sortable fields', () => {
+         component.sortable = false;
+         component.uiDefinitions = {
+            memory: {
+               sortable: true
+            }
+         };
+         component.jsonSchema = jsonSchema;
+         fixture.detectChanges();
+         let headerItems: HTMLTableHeaderCellElement[] = fixture.nativeElement.querySelectorAll('.st-table__header-item');
+
+         expect(headerItems[0].classList).not.toContain('st-table__header-item--sortable');
+         expect(headerItems[1].classList).toContain('st-table__header-item--sortable');
+         expect(headerItems[2].classList).not.toContain('st-table__header-item--sortable');
+      });
    });
 
    it('if table is filterable, filterable class is only added to filterable fields', () => {
@@ -104,40 +182,40 @@ describe('StTableComponent', () => {
 
       expect(headerItems[0].classList).not.toContain('st-table__header-item--filterable');
       expect(headerItems[1].classList).not.toContain('st-table__header-item--filterable');
-      expect(headerItems[2].classList).not.toContain('st-table__header-item--filterable');
-      expect(headerItems[3].classList).not.toContain('st-table__header-item--filterable');
-      expect(headerItems[4].classList).toContain('st-table__header-item--filterable');
+      expect(headerItems[2].classList).toContain('st-table__header-item--filterable');
    });
 
    it('If fields input is not introduced, it throws an error', () => {
-      component.fields = undefined;
+      component.jsonSchema = undefined;
       try {
          fixture.detectChanges();
-         expect(component.fields).toThrow();
+         expect(component.jsonSchema).toThrow();
       } catch (error) {
-         expect(error.message).toContain('st-table-component: field fields is a required field');
+         expect(error.message).toContain('st-dynamic-table-component: field jsonSchema is a required field');
       }
    });
 
    it('only is displayed an arrow next to header of the field over which current order is applied', () => {
       component.sortable = true;
       component.currentOrder = undefined;
+      component.jsonSchema = jsonSchema;
       fixture.detectChanges();
 
       expect(fixture.nativeElement.querySelector('thead .st-table__order-arrow')).toBeNull();
 
-      component.currentOrder = new Order(fakeFields[3].id, ORDER_TYPE.ASC);
+      component.currentOrder = new Order('home', ORDER_TYPE.ASC);
       fixture.detectChanges();
 
       let headerItems: HTMLTableHeaderCellElement[] = fixture.nativeElement.querySelectorAll('.st-table__header-item');
 
       expect(fixture.nativeElement.querySelectorAll('thead .icon-arrow-up').length).toBe(1);
-      expect(headerItems[3].querySelector('.icon-arrow-up')).not.toBeNull();
-      expect(headerItems[3].classList).toContain('st-table__header-item--sortable');
+      expect(headerItems[2].querySelector('.icon-arrow-up')).not.toBeNull();
+      expect(headerItems[2].classList).toContain('st-table__header-item--sortable');
    });
 
    it('only is displayed an arrow next to header of the field over which current filter is configured', () => {
       component.filterable = true;
+      component.currentOrder = new Order('home', ORDER_TYPE.DESC);
       fixture.detectChanges();
       expect(fixture.nativeElement.querySelector('thead .st-table__filter-arrow')).toBeDefined();
       let headerItems: HTMLTableHeaderCellElement[] = fixture.nativeElement.querySelectorAll('.st-table__header-item');
@@ -147,103 +225,42 @@ describe('StTableComponent', () => {
       expect(headerItems[0].classList).not.toContain('st-table__header-item--filterable');
       expect(headerItems[1].querySelector('.icon-arrow4_down')).toBeNull();
       expect(headerItems[1].classList).not.toContain('st-table__header-item--filterable');
-      expect(headerItems[2].querySelector('.icon-arrow4_down')).toBeNull();
-      expect(headerItems[2].classList).not.toContain('st-table__header-item--filterable');
-      expect(headerItems[3].querySelector('.icon-arrow4_down')).toBeNull();
-      expect(headerItems[3].classList).not.toContain('st-table__header-item--filterable');
-      expect(headerItems[4].querySelector('.icon-arrow4_down')).not.toBeNull();
-      expect(headerItems[4].classList).toContain('st-table__header-item--filterable');
+      expect(headerItems[2].querySelector('.icon-arrow4_down')).not.toBeNull();
+      expect(headerItems[2].classList).toContain('st-table__header-item--filterable');
    });
 
-   describe('Should return the class name for header items according to the current order and direction', () => {
-      beforeEach(() => {
-         fixture.detectChanges();
-      });
-
-      it('if current order is not defined yet, no one arrow is displayed', () => {
-         component.currentOrder = undefined;
-         fixture.detectChanges();
-
-         expect(fixture.nativeElement.querySelector('thead st-table__order-arrow')).toBeNull();
-      });
-
-      it('if table is sort by the field but not in ascending direction, it returns icon-arrow-down', () => {
-         component.currentOrder = new Order(fakeFields[0].id, ORDER_TYPE.DESC);
-
-         expect(component.getHeaderItemClass(fakeFields[0])).toBe('icon-arrow-down');
-      });
-
-      it('if table is sort in ascending direction but not by the introduced field, it returns icon-arrow-down', () => {
-         component.currentOrder = new Order(fakeFields[1].id, ORDER_TYPE.ASC);
-
-         expect(component.getHeaderItemClass(fakeFields[0])).toBe('icon-arrow-down');
-      });
-
-      it('if table is sort by that field and in ascending direction, it returns icon-arrow-up', () => {
-         component.currentOrder = new Order(fakeFields[0].id, ORDER_TYPE.ASC);
-
-         expect(component.getHeaderItemClass(fakeFields[0])).toBe('icon-arrow-up');
-      });
-
-   });
 
    describe('When user clicks on a field in the table header, order is changed', () => {
       beforeEach(() => {
          spyOn(component.changeOrder, 'emit');
          fixture.detectChanges();
       });
-      it('if field is different to the current order`s one, current order is changed to the selected field and in direction ASC', () => {
-         component.currentOrder = new Order(fakeFields[0].id, ORDER_TYPE.ASC);
 
-         let headerItem: HTMLTableHeaderCellElement = fixture.nativeElement.querySelectorAll('.st-table__header-item label')[fakeFields.length - 1];
+      it('if field is different to the current order`s one, current order is changed to the selected field and in direction ASC', () => {
+         component.currentOrder = new Order('cores', ORDER_TYPE.ASC);
+
+         let headerItem: HTMLTableHeaderCellElement = fixture.nativeElement.querySelectorAll('.st-table__header-item label')[2];
          headerItem.click();
-         fixture.changeDetectorRef.markForCheck();
          fixture.detectChanges();
 
          expect(component.changeOrder.emit).toHaveBeenCalled();
          const expectedOrder: Order = (<jasmine.Spy> component.changeOrder.emit).calls.mostRecent().args[0];
-         expect(expectedOrder.orderBy).toBe(fakeFields[fakeFields.length - 1].id);
-         expect(expectedOrder.type).toBe(ORDER_TYPE.ASC);
-      });
-
-      it('if field is the same to the current order`s one, only order direction is changed', () => {
-         // ascent sorting
-         component.currentOrder = new Order(fakeFields[0].id, ORDER_TYPE.ASC);
-         let headerItem: HTMLTableHeaderCellElement = fixture.nativeElement.querySelectorAll('.st-table__header-item label')[0];
-         headerItem.click();
-         fixture.detectChanges();
-
-         expect(component.changeOrder.emit).toHaveBeenCalled();
-         let expectedOrder: Order = (<jasmine.Spy> component.changeOrder.emit).calls.mostRecent().args[0];
-         expect(expectedOrder.orderBy).toBe(fakeFields[0].id);
-         expect(expectedOrder.type).toBe(ORDER_TYPE.DESC);
-
-         // descent sorting
-         component.currentOrder = new Order(fakeFields[0].id, ORDER_TYPE.DESC);
-         headerItem.click();
-         fixture.detectChanges();
-
-         expectedOrder = (<jasmine.Spy> component.changeOrder.emit).calls.mostRecent().args[0];
-         expect(component.changeOrder.emit).toHaveBeenCalled();
-         expect(expectedOrder.orderBy).toBe(fakeFields[0].id);
+         expect(expectedOrder.orderBy).toBe('home');
          expect(expectedOrder.type).toBe(ORDER_TYPE.ASC);
       });
 
       it('if field is undefined, order is not changed', () => {
-         component.currentOrder = new Order(fakeFields[1].id, ORDER_TYPE.DESC);
-
          component.onChangeOrder(undefined);
          fixture.detectChanges();
 
-         expect(component.currentOrder.orderBy).toBe(fakeFields[1].id);
-         expect(component.currentOrder.type).toBe(ORDER_TYPE.DESC);
-         expect(component.changeOrder.emit).not.toHaveBeenCalled();
+         expect(component.changeOrder.emit).not.toHaveBeenCalledWith(undefined);
       });
 
       it('should stand up the field`s header which table is sorted by', () => {
-         component.currentOrder = new Order(fakeFields[fakeFields.length - 1].id, ORDER_TYPE.ASC);
-         let headerItem: HTMLTableHeaderCellElement = fixture.nativeElement.querySelectorAll('.st-table__header-item label')[fakeFields.length - 1];
+         component.sortable = true;
+         component.currentOrder = new Order('home', ORDER_TYPE.ASC);
          fixture.detectChanges();
+         let headerItem: HTMLTableHeaderCellElement = fixture.nativeElement.querySelectorAll('.st-table__header-item label')[2];
 
          expect(headerItem.parentElement.classList).toContain('st-table__header-item--selected');
       });
@@ -259,30 +276,29 @@ describe('StTableComponent', () => {
 
       it('Should show a menu where to apply filters', () => {
          let headerItem: HTMLTableHeaderCellElement = fixture.nativeElement.querySelectorAll('.st-table__header-item');
-         headerItem[4].querySelector('.st-table__filter-arrow').click();
-         fixture.changeDetectorRef.markForCheck();
+         headerItem[2].querySelector('.st-table__filter-arrow').click();
          fixture.detectChanges();
          let popover = fixture.nativeElement.querySelectorAll('.st-table__popover-content')[0];
 
-         if (!component.templateContentFilter) {
-            expect(popover).not.toBeNull();
-            expect(popover.querySelector('st-checkbox')).not.toBeNull();
-         }
+         expect(popover).not.toBeNull();
+         expect(popover.querySelector('st-checkbox')).not.toBeNull();
       });
 
       it('if not has custom template and select an option, clicking on button should emit eventEmitter with selected filters', () => {
          let headerItem: HTMLTableHeaderCellElement = fixture.nativeElement.querySelectorAll('.st-table__header-item');
-         headerItem[4].querySelector('.st-table__filter-arrow').click();
+         headerItem[2].querySelector('.st-table__filter-arrow').click();
          fixture.detectChanges();
          let popover = fixture.nativeElement.querySelectorAll('.st-table__popover-content')[0];
          popover.querySelector('.st-table__popover-button').click();
          fixture.detectChanges();
-         expect(headerItem[4].querySelector('.icon-arrow4_down')).not.toBeNull();
+         expect(headerItem[2].querySelector('.icon-arrow4_down')).not.toBeNull();
 
+         (<jasmine.Spy> component.selectFilters.emit).calls.reset();
          popover.querySelector('st-checkbox').querySelector('input').click();
          popover.querySelector('.st-table__popover-button').click();
-
-         expect(component.selectFilters.emit).toHaveBeenCalledWith([fakeFields[4]]);
+         const expectedHeader: StTableHeader = _cloneDeep(component.fields[2]);
+         expectedHeader.filters.filterConfig = [{ id: '/root', name: '/root', selected: true }];
+         expect(component.selectFilters.emit).toHaveBeenCalledWith([expectedHeader]);
       });
    });
 
@@ -338,18 +354,16 @@ describe('StTableComponent', () => {
 
             expect(selectedAllCheckbox.checked).toBeFalsy();
          });
-
       });
    });
 
    it('only if hasHoverMenu is true, an empty cell is created after the header cells', () => {
       component.hasHoverMenu = true;
-      component.fields = fakeFields;
-
       fixture.detectChanges();
+
       let headerCells: HTMLTableHeaderCellElement[] = fixture.nativeElement.querySelectorAll('th');
 
-      expect(headerCells.length).toBe(fakeFields.length + 1);
+      expect(headerCells.length).toBe(component.fields.length + 1);
       expect(headerCells[headerCells.length - 1].innerHTML).toEqual('');
 
       component.hasHoverMenu = false;
@@ -357,11 +371,13 @@ describe('StTableComponent', () => {
       fixture.detectChanges();
       fixture.changeDetectorRef.markForCheck();
 
-      expect(fixture.nativeElement.querySelectorAll('th').length).toBe(fakeFields.length);
+      expect(fixture.nativeElement.querySelectorAll('th').length).toBe(component.fields.length);
    });
 
    describe('it should fix its header in order to be displayed although page scrolling', () => {
       it('if input "fixedHeader" is not specified, header is not fixed', () => {
+         fixture.detectChanges();
+
          expect(fixture.nativeElement.querySelector('table').classList).not.toContain('st-table--fixed-header');
       });
 
@@ -381,11 +397,12 @@ describe('StTableComponent', () => {
    });
 
    it('Custom classes can be added to the table', () => {
+      fixture.detectChanges();
+
       let fakeClass = 'separated-rows';
       expect(fixture.nativeElement.querySelector('table').classList).not.toContain(fakeClass);
 
       component.customClasses = fakeClass;
-
       fixture.detectChanges();
 
       expect(fixture.nativeElement.querySelector('table').classList).toContain('separated-rows');