+++ projects/egeo/src/lib/st-dynamic-table/st-dynamic-table.model.ts
@@ -1,0 +1,18 @@
+import { StTableHeader } from '../';
+
+
+export interface StDynamicTableHeader extends StTableHeader {
+   reference: string;
+}
+
+export interface StDynamicTableUISpecification {
+   sortable?: boolean;
+   styles?: {
+      [key: string]: string;
+   };
+}
+
+
+export interface StDynamicTableUserInterface {
+   [key: string]: StDynamicTableUISpecification;
+}