+++ projects/egeo/src/lib/st-table/shared/st-table-row/st-table-row.component.ts
@@ -8,10 +8,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import {
-   ChangeDetectionStrategy, Component, HostListener, Input, OnInit, ViewChild, HostBinding,
-   ElementRef
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, HostBinding, HostListener, Input, ViewChild } from '@angular/core';
 
 /**
  * @description {Component} [Table Row]
@@ -37,13 +34,13 @@ import {
    changeDetection: ChangeDetectionStrategy.OnPush
 })
 
-export class StTableRowComponent implements OnInit {
+export class StTableRowComponent {
    /** @Input {boolean} [selected=''] It indicates if row is selected or not */
    @Input() selected: boolean;
    /** @Input {boolean} [standUpSelected=true] It indicates if when row is selected, it has to be displayed stood up */
    @Input() standUpSelected: boolean = true;
 
-   @ViewChild('hoverMenu', {static: true}) hoverMenu: any;
+   @ViewChild('hoverMenu', { static: false }) hoverMenu: any;
 
    @HostBinding('class.st-table-row') rootClass: boolean = true;
 
@@ -52,18 +49,7 @@ export class StTableRowComponent implements OnInit {
       return this.selected && this.standUpSelected;
    }
 
-
    public showHoverMenu: boolean = false;
-
-   constructor(private elementRef: ElementRef) {
-
-   }
-
-   ngOnInit(): void {
-      if (this.hoverMenu.nativeElement.children.length === 0) {
-         this.elementRef.nativeElement.removeChild(this.hoverMenu.nativeElement);
-      }
-   }
 
    @HostListener('mouseover') onShowHoverMenu(): void {
       this.showHoverMenu = true;